// src/pages/auth/types.tsx
export const AuthProvider = {
    LOCAL: 'LOCAL',
    GOOGLE: 'GOOGLE',
} as const;

export type AuthProvider = typeof AuthProvider[keyof typeof AuthProvider];

export const Roles = {
    ADMIN: 'ADMIN',
    TEACHER: 'TEACHER',
    STUDENT: 'STUDENT',
    SUPERADMIN: 'SUPERADMIN',
} as const;

export type Roles = typeof Roles[keyof typeof Roles];

export const TeacherSpecification = {
    IELTS: 'IELTS',
    SPEAKING: 'SPEAKING',
    GRAMMAR: 'GRAMMAR',
    BUSINESS: 'BUSINESS',
    KIDS: 'KIDS',
} as const;

export type TeacherSpecification = typeof TeacherSpecification[keyof typeof TeacherSpecification];

export const LessonStatus = {
    AVAILABLE: 'AVAILABLE',
    BOOKED: 'BOOKED',
    COMPLETED: 'COMPLETED',
    CANCELLED: 'CANCELLED',
} as const;

export type LessonStatus = typeof LessonStatus[keyof typeof LessonStatus];

export interface Teacher {
    id: string;
    email: string;
    phoneNumber: string;
    fullName: string;
    cardNumber?: string;
    isActive?: boolean;
    authProvider?: AuthProvider;
    isDelete?: boolean;
    isComplete?: boolean;
    role?: Roles;
    specification?: TeacherSpecification;
    level?: string;
    description?: string;
    hourPrice?: number;
    portfolioLink?: string;
    imageUrl?: string;
    rating: number;
    experience?: string;
    createdAt?: string;
    updatedAt?: string;
    // Relations
    lessonsCount?: number;
}

export interface Student {
    id: string;
    lastName: string;
    firstName: string;
    phoneNumber?: string;
    role: Roles;
    tgId?: string;
    tgUsername?: string;
    isBlocked: boolean;
    blockedAt?: Date;
    blockedReason?: string;
    email?: string;
    createdAt: string;
    updatedAt: string;
}

export interface Lesson {
    id: string;
    name: string;
    startTime: string;
    endTime: string;
    teacherId: string;
    studentId?: string;
    googleMeetUrl?: string;
    status: LessonStatus;
    googleEventId?: string;
    price: number;
    isPaid: boolean;
    bookedAt?: string;
    completedAt?: string;
    createdAt: string;
    updatedAt: string;
    teacher?: Teacher;
    student?: Student;
}